requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
Pillow>=8.3.1
numpy>=1.21.0
torch>=1.9.0  # Se estiver usando PyTorch no megadetector
torchvision>=0.10.0  # Se estiver usando PyTorch
opencv-python>=4.5.0  # Possivelmente necessário para o megadetector
shutil-py>=1.0.0  # Normalmente já vem com Python, mas incluído por garantia
python-dotenv>=0.19.0  # Útil para gerenciamento de variáveis de ambiente